API consumption from WebAP
With this commit, WebAP can be consumed from webpages when
the extension is installed and configured.
The first ever message was posted using this method by
directly consuming WebAP from the console at an arbitrary
webpage.
For historical reason, I'll register the webpage here:
https://github.com/broofa/mime
As there was a typo on the message, I have deleted it.
So, here is a link to the second ever message posted
with WebAP :D :
https://blob.cat/notice/AS12v9WLNjNZtcWhEm

My original plan was to consume the APIs from the
content script, but that was not possible. I had
CORS issues during testing, so I moved the consumption
to the background service worker.